refactor: separate middleware
